Add pundit

+++ app/controllers/sources_controller.rb
@@ -1,17 +1,18 @@
 # controller for Sources
 class SourcesController < ApplicationController
-  before_action :authenticate_user!
 
   def index
-    @sources = current_user.sources
+    @sources = policy_scope(Source)
   end
 
   def choose_sources
     @sources = Source.all
     @locale = params[:locale]
+    authorize Source
   end
 
-  def add_sources_to_user
+  def update_sources_for_user
+    authorize Source
     current_user.sources.delete_all
     if params[:source_ids].present?
       ids = params[:source_ids].map(&:to_i)
@@ -24,7 +25,8 @@ class SourcesController < ApplicationController
 
   def edit_sources_for_user
     @sources = Source.all
-    @user_sources = current_user.sources
+    authorize Source
+    @user_sources = policy_scope(Source)
     @locale = params[:locale]
   end
 end
